requirements.txt
selenium
webdriver-manager
pandas
pyyaml